thread safe and dump access in the same function
