+++ core/src/drivers/plugins/python/modbus_slave_plugin/modbus_slave_config.ini
