+++ plugins.conf
@@ -1,4 +1,4 @@
 # Plugin configuration file
 # Format: name,path,enabled,type,config_path
-# modbus_slave,../core/src/drivers/plugins/python/simple_modbus.py,1,0,./modbus_slave_config.ini
-example_plugin,./core/src/drivers/examples/example_python_plugin.py,1,0,./modbus_slave_config.ini
+modbus_slave,./core/src/drivers/plugins/python/modbus_slave_plugin/simple_modbus.py,1,0,./modbus_slave_config.ini
+example_plugin,./core/src/drivers/examples/example_python_plugin.py,1,0,./example_plugin_config.ini
